backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
anyio==4.11.0
asgiref==3.9.1
attrs==25.3.0
black==25.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.1
Django==5.2.6
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
firebase_admin==7.1.0
flake8==7.3.0
frozenlist==1.7.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.75.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
mccabe==0.7.0
msgpack==1.1.1
multidict==6.6.4
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
PyJWT==2.10.1
pytokens==0.1.10
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
sqlparse==0.5.3
stripe==12.5.1
twilio==9.8.2
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.20.1
